+++ R/ShinyGUIClass.R
@@ -65,7 +65,7 @@ ShinyGUI <- setRefClass("ShinyGUI",
                               if(length(shinyGUI)){
                                   dnames <- names(shinyGUI)
                                   gnames <- c('title', 'authors', 'instructions', 'firstpage', 'demographics',
-                                              'demographics_inputIDs', 'max_time', 'temp_file', 'resume_file',
+                                              'demographics_inputIDs', 'max_time', 'temp_file', 
                                               'lastpage', 'css', 'stem_dims', 'forced_choice', 'stem_locations',
                                               'begin_message')
                                   if(!all(dnames %in% gnames))