+++ tests/GUI-tests/GUI-tests.R
@@ -24,6 +24,12 @@ lastpagefun <- function(person){
 } 
 results2 <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE, lastpage=lastpagefun))
 
+# save and resume temp file
+mirtCAT(df = df, shinyGUI = list(temp_file = 'thisfile.rds'))
+
+mirtCAT(df = df, shinyGUI = list(temp_file = 'thisfile.rds')) #stop early
+mirtCAT(df = df, shinyGUI = list(temp_file = 'thisfile.rds')) #this resumes
+
 ## two step hosting
 my_fun <- function(person) cat('Hello world\n')
 mirtCAT_preamble(df, final_fun = my_fun)